pyproject.toml
[project]
name = "code-jangler"
version = "0.1.0"
description = "An agentic system that splits PRs into a stack of smaller, easier-to-review PRs"
requires-python = ">=3.10"
dependencies = [
    "openai-agents",
    "openai",
    "streamlit",
    "uuid",
    "pydantic",
    "python-dotenv",
    "asyncio",
    "openai-agents-mcp>=0.0.6",
    "watchdog>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 88
target-version = "py38"
